Experimented with separate ViewModel
Experimented with using separate ViewModel for shared logic.
