+++ src/Chirp.Web/Pages/Shared/_ShareBar.cshtml
@@ -1,12 +1,13 @@
-﻿@model Chirp.Web.Pages.PublicModel
+﻿@model Chirp.Web.Pages.Shared.ShareCheepViewModel
 
 <div class="cheepbox">
     <h3>What's on your mind @(User.Identity.Name)?</h3>
-    <form asp-page="SubmitCheep">
+    <form asp-page="/Shared/ShareCheepViewModel">
+    @* <form asp-page="SubmitCheep"> *@
         <input style="float: left" type="text" asp-for="CheepText">
         <input type="submit" value="Share">
         <div style="clear: both;"></div> <!-- Clears float -->
         <span asp-validation-for="CheepText" class="text-danger"></span>
-        <br />
+        <br/>
     </form>
 </div>
