+++ src/Chirp.Web/Pages/UserTimeline.cshtml
@@ -17,13 +17,8 @@
         <h2> @authorName's Timeline </h2>
     </div>
     @if(User.Identity.IsAuthenticated){
-    <div class="cheepbox">
-        <h3>What's on your mind @(User.Identity.Name)?</h3>
-        <form method="post">
-            <input style="float: left" type="text" asp-for="Text">
-            <input type="submit" value="Share">
-        </form>
-    </div>
+        @* Render the Share Form Partial View *@
+        @await Html.PartialAsync("_ShareBar", Model.ShareCheepView)
     }
     @if (Model.Cheeps.Any())
     {
